Move PauliEvolutionGate label tests into test_evolution_gate.py and replace assertIn with exact label checks
